added API for subscribing wallets manually
